requirements.txt
pytest==4.2.0
selenium==3.141.0
PyYAML==3.13
